refactor utils and racf_functions
